Dump results to pickle offline to prevent server timeout
As Heroku has a 60 seconds limitation for app boot time,
grading all students will cause server crashing.
In this commit, the  results would be dumped to results/results.pickle
offline to prevent server timeout.
This is just a workaround. If there is a way/service to solve this
problem, the results.pickle could be removed.