Dump results to pickle offline to prevent server timeout
As Heroku has a 60 seconds limitation for app boot time,
grading all students will cause server crashing.
In this commit, the  results would be dumped to results/results.pickle
offline to prevent server timeout.
This is just a workaround. If there is a way/service to solve this
problem, the results.pickle could be removed.


Former-commit-id: 1db0d7eb420d718fc156814649ab6539f60c0421 [formerly 1905486c063667042906fa3e95524e02cf5330d2]
Former-commit-id: 908533740ee9f0337513813fe7ddac863e14aa98